change get_members_for_excluding, rename request_sevice
