+++ src/core/services/request_service.py
@@ -61,13 +61,3 @@ class RequestService:
         request.status = Request.Status.DECLINED
         await self.__request_repository.update(request_id, request)
         return RequestResponse.parse_from(request)
-
-    async def exclude_members(self, members: list[Member], bot: Application.bot) -> None:
-        """Исключает участников смены.
-
-        Аргументы:
-            members (list[Member]): список участников подлежащих исключению
-        """
-        await self.__member_repository.update_status_to_exclude(members)
-        for member in members:
-            await self.__telegram_bot(bot).notify_excluded_member(member.user)
